converter/requirements.txt
fastapi
uvicorn
pydantic
requests
python-multipart
mammoth
beautifulsoup4
lxml
python-dotenv
anthropic
